SelectControl: Move types
